+++ src/ampio_mqtt/__init__.py
@@ -41,4 +41,4 @@ __all__ = [
     "module_model",
 ]
 
-__version__ = "0.12.0"
+__version__ = "0.13.0"